Require per-method Lodash packages
I tried a few permutations and eventually the TypeScript-specific
`import some = require('lodash.some')`, which works fine for the ES5
build, but not ES2015 per:

https://github.com/Microsoft/TypeScript/issues/15426.

Falling back to a standard `require` was the only way I could make this
work (thus far).

However, this means `index.es2015.js` is no longer an ES module. Lmk if
there's a better way of this.

Closes #36.